requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiohttp=3.9.3=pypi_0
aiohttp-cors=0.7.0=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
brotli-python=1.0.9=py310h6a678d5_7
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=py310h06a4308_0
chardet=4.0.0=py310h06a4308_1003
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
colorful=0.5.6=pypi_0
compiler-gym=0.2.5=pypi_0
contourpy=1.2.0=pypi_0
cpuonly=2.0=0
cycler=0.12.1=pypi_0
deprecated=1.2.14=pypi_0
distlib=0.3.8=pypi_0
docker=7.0.0=pypi_0
fasteners=0.19=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
fonttools=4.50.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-api-core=2.17.1=pypi_0
google-auth=2.28.2=pypi_0
googleapis-common-protos=1.63.0=pypi_0
gpustat=1.1.1=pypi_0
grpcio=1.43.0=pypi_0
gym=0.21.0=pypi_0
humanize=4.9.0=pypi_0
idna=3.6=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py310h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
loop-tool-py=0.0.7=pypi_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py310h5eee18b_0
matplotlib=3.8.3=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
msgpack=1.0.8=pypi_0
multidict=6.0.5=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=pypi_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
nvidia-ml-py=12.535.133=pypi_0
opencensus=0.11.4=pypi_0
opencensus-context=0.1.3=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
packaging=24.0=pypi_0
pandas=2.2.1=pypi_0
pillow=10.2.0=py310h5eee18b_0
pip=24.0=pypi_0
platformdirs=4.2.0=pypi_0
prometheus-client=0.20.0=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.8=pypi_0
py-spy=0.3.14=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pybind11=2.11.1=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py310h06a4308_0
pysqlite3=0.5.2=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.2.1=py3.10_cpu_0
pytorch-mutex=1.0=cpu
pytz=2024.1=pypi_0
pyyaml=6.0.1=py310h5eee18b_0
ray=2.9.3=pypi_0
readline=8.2=h5eee18b_0
referencing=0.33.0=pypi_0
requests=2.31.0=py310h06a4308_1
rpds-py=0.18.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.13.0=pypi_0
setuptools=66.0.0=pypi_0
six=1.16.0=pypi_0
smart-open=7.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=py310h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.4.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.2.1=py310_cpu
torchvision=0.17.1=py310_cpu
tqdm=4.66.2=pypi_0
typing-extensions=4.10.0=pypi_0
typing_extensions=4.9.0=py310h06a4308_1
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
virtualenv=20.25.1=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.41.2=py310h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
